Renamed lib/pacts to lib/pact.
