+++ lib/commands/utils.js
@@ -167,28 +167,6 @@ function getAuthOptionsFromArgs(...args) {
   return undefined;
 }
 
-export function getAuthOptionsFromBasicAuthHeader(authHeader) {
-  if (
-    !authHeader ||
-    !_.isString(authHeader) ||
-    !authHeader.startsWith("Basic ")
-  ) {
-    return undefined;
-  }
-
-  const base64Credentials = authHeader.slice("Basic ".length);
-  const credentials = Buffer.from(base64Credentials, "base64").toString(
-    "utf-8"
-  );
-
-  const components = credentials.split(":");
-  if (!components || components.length < 2) {
-    return undefined;
-  }
-
-  return { user: components[0], password: components.slice(1).join(":") };
-}
-
 export function persistAuth(auth) {
   const win = cy.state("window");
   if (auth) {
