+++ test/cypress/e2e/c8ypact-runner.cy.ts
@@ -1,4 +1,4 @@
-import { C8yDefaultPactRunner } from "@lib/pacts/runner";
+import { C8yDefaultPactRunner } from "@lib/pact/runner";
 import { C8yDefaultPactRecord } from "@shared/c8ypact";
 
 const { _ } = Cypress;
